18.12.17 V1

+++ bot.js
@@ -6,13 +6,14 @@ bot.on("ready", () => {
 });
 
 bot.on("message", message => {
+  if (message.author.bot) return;
   if(message.content.toLowerCase().indexOf('go in') > -1) {
     message.delete();
     message.channel.send(`General Chat is not the place. Please use the private raid room.`);
   }
   else if(message.content.toLowerCase().indexOf('nest list') > -1) {
     message.delete();
-    message.channel.send(`Please visit #bt-updates for information on the current nests!`);
+    message.channel.send(`Please visit \#bt-go-updates for information on the current nests!`);
   }
   else if(message.content.toLowerCase().indexOf('spoofing') > -1) {
     message.delete();
